+++ src/main/kotlin/io/github/janbarari/gradle/IncompatibleVersionException.kt
@@ -20,15 +20,13 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package io.github.janbarari.gradle.analytics
-
-import io.github.janbarari.gradle.analytics.GradleAnalyticsPlugin.Companion.PLUGIN_NAME
-import io.github.janbarari.gradle.ExcludeJacocoGenerated
+package io.github.janbarari.gradle
 
 @ExcludeJacocoGenerated
 class IncompatibleVersionException(
+    private val title: String,
     private val minimumRequiredVersion: String
 ) : Throwable() {
-    override val message: String get() = "$PLUGIN_NAME is compatible with" +
-            " gradle version $minimumRequiredVersion and above."
+    override val message: String
+        get() = "$title is compatible with Gradle version $minimumRequiredVersion and above."
 }
